move encoding.ts to common for #79275
