+++ src/vs/base/test/common/encoding/fixtures/some.cp1252.txt
