+++ src/vs/base/test/common/encoding/fixtures/some.json.png
