+++ src/vs/base/test/common/encoding/fixtures/some.png.txt
