+++ src/vs/base/test/common/encoding/fixtures/some.qwoff.txt
