+++ src/vs/base/test/common/encoding/fixtures/some.shiftjis.txt
