+++ src/vs/base/test/common/encoding/fixtures/some.xml.png
