+++ src/vs/base/test/common/encoding/fixtures/some_ansi.css
