+++ src/vs/base/test/common/encoding/fixtures/some_utf16le.css
