+++ src/vs/base/test/common/encoding/fixtures/some_utf8.css
