+++ src/vs/base/test/common/encoding/fixtures/utf16_be_nobom.txt
