+++ src/vs/base/test/common/encoding/fixtures/utf16_le_nobom.txt
